Work in Progress on css for quiz game
