+++ src/main/java/com/example/geriadur/constants/WordClassEnum.java
@@ -1,5 +1,5 @@
 package com.example.geriadur.constants;
 
 public enum WordClassEnum {
-    WN, WV, WADJ, WADV, WART, WPRN, WPREP, WCONJ, SUF
+    WN, WV, WADJ, WADV, WART, WPRN, WPREP, WCONJ, WAF
 }
