+++ src/main/java/com/example/geriadur/dto/CreateEtymo.java
@@ -1,0 +1,33 @@
+package com.example.geriadur.dto;
+
+import com.example.geriadur.domain.LiteralTranslation;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class CreateEtymo {
+
+
+    private String currentName;
+
+    private String etymoName;
+
+    private List<String> lexemes;
+
+    private String descrFr;
+
+    private String descrEng;
+
+    private Long wordTheme;
+
+    private Long semanticField;
+
+    private LiteralTranslation litTrans;
+
+}
