+++ src/main/java/com/example/geriadur/repositories/EtymonNameRepository.java
@@ -12,6 +12,7 @@ import java.util.Set;
 @Repository
 public interface EtymonNameRepository extends JpaRepository<EtymonName, Long> {
     Optional<EtymonName> findEtymonNameByEtymonId(Long id);
+    Optional<EtymonName> findEtymonNameByCurrentName(String currentName);
     @Query("FROM EtymonName g where g.wordTheme = :word_theme ORDER BY RAND() LIMIT 15")
     Set<EtymonName> findEtymonNamesByWordTheme(@Param("word_theme")int wordTheme);
 }
