+++ src/main/java/com/example/geriadur/util/DataUtil.java
@@ -6,6 +6,9 @@ import com.example.geriadur.domain.consultation.Lexeme;
 import com.example.geriadur.domain.SemanticField;
 import com.example.geriadur.domain.consultation.Quote;
 import com.example.geriadur.domain.consultation.Source;
+import com.example.geriadur.dto.CreateEtymo;
+import com.example.geriadur.dto.CreateLexeme;
+import com.example.geriadur.dto.CreateSource;
 import com.example.geriadur.dto.UserRegistrationDto;
 import com.example.geriadur.repositories.*;
 import com.example.geriadur.service.user.UserServiceImpl;
@@ -41,10 +44,11 @@ public class DataUtil {
     private LiteralTranslationRepository literalTranslationRepository;
     @Autowired
     private QuoteRepository quoteRepository;
-    private List<Lexeme> lexemesInit = new ArrayList<>();
-    private List<EtymonName> etymonNamesInit = new ArrayList<>();
+    private List<CreateLexeme> lexemesInit = new ArrayList<>();
+    private List<CreateEtymo> etymonNamesInit = new ArrayList<>();
     private List<Source> sourcesInit = new ArrayList<>();
     private List<Quote> quotesInit = new ArrayList<>();
+
     @PostConstruct
     public void InjectionData() throws IOException {
         readJsonData("lexemesInit");
@@ -67,48 +71,55 @@ public class DataUtil {
         semanticField4.setSemanticFieldNameFr("religion");
         ArrayList<LiteralTranslation> literalTranslations = new ArrayList<>();
         //ArrayList<EtymonName> etymonNames = new ArrayList<>();
-        for (EtymonName etymonName : etymonNamesInit) {
-            literalTranslations.add(etymonName.getLitTrans());
-            etymonName.setSemanticField(semanticField);
-        }
-        lexemeRepository.saveAll(lexemesInit);
-        literalTranslationRepository.saveAll(literalTranslations);
         semanticFieldRepository.save(semanticField);
         semanticFieldRepository.save(semanticField2);
         semanticFieldRepository.save(semanticField3);
         semanticFieldRepository.save(semanticField4);
-        etymonNameRepository.saveAll(etymonNamesInit);
+
         sourceRepository.saveAll(sourcesInit);
         quoteRepository.saveAll(quotesInit);
+        ArrayList<EtymonName> etymonNames = new ArrayList<>();
+        for (CreateEtymo createEtymo : etymonNamesInit) {
+            literalTranslations.add(createEtymo.getLitTrans());
+            EtymonName etymonName = new EtymonName();
+            etymonName.setEtymoName(createEtymo.getEtymoName());
+            etymonName.setCurrentName(createEtymo.getCurrentName());
+            etymonName.setEtymoName(createEtymo.getEtymoName());
+            etymonName.setWordTheme(createEtymo.getWordTheme());
+            etymonName.setLitTrans(createEtymo.getLitTrans());
+            etymonName.setDescrFr(createEtymo.getDescrFr());
+            etymonName.setDescrEng(createEtymo.getDescrEng());
+            /* semanticfield is not define in etymonsInit.json yet*/
+            etymonName.setSemanticField(semanticFieldRepository.findById(1L).get());
+            etymonNames.add(etymonName);
+        }
+        literalTranslationRepository.saveAll(literalTranslations);
+        etymonNameRepository.saveAll(etymonNames);
+        ArrayList<Lexeme> lexemes = new ArrayList<>();
+        for (CreateLexeme createLexeme : lexemesInit) {
+            Lexeme lexeme = new Lexeme();
+            lexeme.setLexemeName(createLexeme.getLexemeName());
+            lexeme.setLexemeLanguage(createLexeme.getLexemeLanguage());
+            lexeme.setGender(createLexeme.getGender());
+            lexeme.setWordClass(createLexeme.getWordClass());
+            lexeme.setDescrEng(createLexeme.getDescrEng());
+            lexeme.setDescrFr(createLexeme.getDescrFr());
+            lexeme.setReferenceWordsEng(createLexeme.getReferenceWordsEng());
+            lexeme.setReferenceWordsFr(createLexeme.getReferenceWordsFr());
+            lexeme.setPhonetic(createLexeme.getPhonetic());
+            lexemes.add(lexeme);
+        }
+
+        lexemeRepository.saveAll(lexemes);
         setLexemeQuoteLink(quoteRepository.findByQuoteId(1L).get(), "bailh");
         setLexemeQuoteLink(quoteRepository.findByQuoteId(2L).get(), "seno");
         setSourceQuoteLink(quoteRepository.findByQuoteId(1L).get(), 2L);//Dictionaire Rostrenen
         setSourceQuoteLink(quoteRepository.findByQuoteId(2L).get(), 6L);//Dictionnaire gaulois delamarre
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(1L).get(), Arrays.asList("uer", "kinget", "rix"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(2L).get(), Arrays.asList("kara", "uos", "os"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(3L).get(), Arrays.asList("kom", "bhiyo", "os"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(4L).get(), Arrays.asList("aku", "kuno"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(5L).get(), Arrays.asList("kuno", "belin", "os"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(6L).get(), Arrays.asList("talu", "iestin"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(7L).get(), Arrays.asList("bran", "os"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(8L).get(), Arrays.asList("ambi", "catu", "os"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(9L).get(), Arrays.asList("ambio", "rix"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(10L).get(), Arrays.asList("uer", "tigern"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(11L).get(), Arrays.asList("argntom", "onios"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(12L).get(), Arrays.asList("ari", "uid"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(13L).get(), Arrays.asList("kuno", "maros"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(14L).get(), Arrays.asList("teuta", "ualos"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(15L).get(), Arrays.asList("korios", "soldurio"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(16L).get(), Arrays.asList("ostimios"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(17L).get(), Arrays.asList("ad", "treb", "tegos"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(18L).get(), Arrays.asList("karnon"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(19L).get(), Arrays.asList("belin", "uak"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(20L).get(), Arrays.asList("keno", "manos"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(21L).get(), Arrays.asList("leng"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(22L).get(), Arrays.asList("sec","gwanu"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(23L).get(), Arrays.asList("allo","brog"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(24L).get(), Arrays.asList("ari","uerno"));
-        setLexemeEtymonLink(etymonNameRepository.findEtymonNameByEtymonId(25L).get(), Arrays.asList("ambi","ianos"));
+
+        for (CreateEtymo etymonName : etymonNamesInit) {
+            setLexemeEtymonLink(etymonNameRepository.findEtymonNameByCurrentName(etymonName.getCurrentName()).get(), etymonName.getLexemes());
+        }
+
         userService.save(new UserRegistrationDto("UserAccount", "lastname", "email", "pass", "U", 1));
         userService.save(new UserRegistrationDto("Admin", "lastname", "emailAdmin", "pass", "A", 1));
     }
@@ -119,12 +130,13 @@ public class DataUtil {
     }
 
     public void setLexemeQuoteLink(Quote quote, String lexemeStr) {
-            Optional<Lexeme> lexeme = lexemeRepository.findByLexemeName(lexemeStr);
-            if (lexeme.isPresent()) {
-                quote.setLexemes(Arrays.asList(lexeme.get()));
-                quoteRepository.save(quote);
-            } else throw new IllegalArgumentException("the lexeme: " + lexemeStr  + " doesn't exist in DB.");
-        }
+        Optional<Lexeme> lexeme = lexemeRepository.findByLexemeName(lexemeStr);
+        if (lexeme.isPresent()) {
+            quote.setLexemes(Arrays.asList(lexeme.get()));
+            quoteRepository.save(quote);
+        } else throw new IllegalArgumentException("the lexeme: " + lexemeStr + " doesn't exist in DB.");
+    }
+
     public void setLexemeEtymonLink(EtymonName etymonName, List<String> lexemesString) {
         Map<Integer, Lexeme> lexemes = new HashMap<>();
         for (int i = 0; i < lexemesString.size(); ++i) {
