+++ src/main/resources/static/css/main.css
@@ -1,16 +1,29 @@
-h2{
-    font-family: sans-serif;
-    font-size: 1.5em;
+h1 {
+    font-family: Calibri, serif;
+    font-size: x-large;
+    padding: 5px;
+    color: burlywood;
     text-transform: uppercase;
+    text-align: center;
 }
 
-strong{
-    font-weight:700;
-    background-color:yellow;
+h2 {
+    font-family: "Arial Black",serif;
+    font-size: xxx-large;
+    color: burlywood;
+    text-align: center;
+    padding: 10px;
 }
-p{
+
+strong {
+    font-weight: 700;
+    background-color: yellow;
+}
+
+p {
     font-family: sans-serif;
 }
+
 /* Tooltip container */
 .langLexeme {
     position: relative;
@@ -57,14 +70,102 @@ p{
     visibility: visible;
     opacity: 1;
 }
+.description{
+    max-width: 500px;
+    font-style: italic;
+    margin: auto;
+    font-size: large;
+    font-weight: normal;
+    color: ghostwhite;
+}
 
-.btn{
-    background: #fff;
-    color:#222;
+.etymoForm{
+    text-align: left;
+    color: ghostwhite;
+    font-size: xx-large;
+    font-weight: 800;
+    padding: 30px 20px 20px 15%;
+    border: none;
+}
+
+.answersForm{
+    display: flex;
+}
+
+.etymoBtn2:hover span {
+    display: block;
+    border: none;
+}
+
+.etymoBtn2:hover {
+    position: relative;
+    transition: all 0.3s;
+    border: none;
+}
+/*
+.etymoBtn span {
+    border: #666 2px solid;
+    padding: 5px 20px 5px 5px;
+    z-index: 10;
+    background: #e3e3e3;
+    margin: 15px;
+    width: 200%;
+    position: absolute;
+    top: 15px;
+    border-radius: 20% 20%;
+    text-align: center;
+    display: none;
+    transition: all 0.3s;
+}*/
+
+.etymoBtn2:hover::after {
+    content: attr(title);
+    z-index: 10;
+    font-size: x-large;
+    font-weight: bold;
+    position: absolute;
+    bottom: 70px;
+    width: 200px;
+    left:-50px;
+    /*margin-left: -5px;*/
+    border-width: 2px;
+    border-style: dashed;
+    border-radius: 4px;
+    border-color: #999;
+
+}
+
+.etymoBtn2 {
+    background: transparent;
+    color: burlywood;
+    font-weight: 700;
+    padding: 50px 5px 15px 5px;
+    margin: 10px;
+    text-align: center;
+    cursor: default;
+    transition: all 0.3s;
+    border: none;
+}
+
+.quiz{
+    padding: 100px 200px 100px 200px;
+}
+
+
+.etymoBtn {
+    background: transparent;
+    padding: 30px;
+    text-align: center;
+}
+
+.btn {
+    background: #555;
+    color: burlywood;
+    font-size: large;
     font-weight: 500;
-    width: 100%;
+    width: 40%;
     border: 1px solid #222;
-    padding: 10px;
+    padding: 20px;
     margin: 10px;
     text-align: left;
     border-radius: 4px;
@@ -72,45 +173,90 @@ p{
     transition: all 0.3s;
 }
 
-.btn.correct{
-    background:#9aeabc;
+.btn.correct {
+    background: #9aeabc;
 }
 
-.btn.incorrect{
+.btn.incorrect {
     background: red;
 }
 
-.btn:hover:not([disabled]){
-    background:#222;
-    color:#fff;
+.btn:hover:not([disabled]) {
+    background: ghostwhite;
+    color: #111;
 }
-.btn:disabled{
+
+.btn:disabled {
     cursor: no-drop;
 }
-.menu-btn{
+
+
+#next-btn {
     background: #001e4d;
-    color:#fff;
+    color: #fff;
     font-weight: 500;
     width: 150px;
     border: 0;
-    padding: 10px;
-    margin: 20px;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-
-#next-btn{
-    background: #001e4d;
-    color:#fff;
-    font-weight: 500;
-    width: 150px;
-    border: 0;
-    padding: 10px;
-    margin: 20px auto 0;
+    padding: 20px;
+    margin: 15px auto 0;
     border-radius: 4px;
     cursor: pointer;
     transition: all 0.3s;
     display: none;
 }
 
+/* footer parameters*/
+
+.menu-btn {
+    background: ghostwhite;
+    color: burlywood;
+    position: relative;
+    left: 100px;
+    font-weight: 500;
+    width: 150px;
+    top:30px;
+    padding: 10px;
+    border-radius: 4px;
+    cursor: pointer;
+}
+body{
+    background: #222222;
+}
+header {
+    width: 100%;
+    background: #555555;
+    position: absolute;
+    top: 0;
+    padding: 20px;
+    height: 100px;
+}
+
+footer {
+    background: #555555;
+    position: absolute;
+    bottom: 0;
+    padding: 20px;
+    width: 100%;
+    height: 150px;
+}
+
+
+/*Functional Styling for tooltip on mobile device*/
+@media (pointer: coarse), (hover: none) {
+    [title] {
+        position: relative;
+        display: flex;
+        justify-content: center;
+    }
+
+    [title]:focus::after {
+        content: attr(title);
+        position: absolute;
+        top: 90%;
+        color: #000;
+        background-color: #fff;
+        border: 1px solid;
+        width: fit-content;
+        padding: 3px;
+    }
+}
