+++ src/main/resources/static/js/configData.js
