+++ src/main/resources/templates/showJson.html
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
     <link rel="stylesheet" href="../css/main.css">
 
-    <script type="text/javascript" th:src="@{/js/script.js}" defer></script>
+    <script type="text/javascript" th:src="@{/js/scriptGame.js}" defer></script>
 </head>
 <body>
 <div id="myData"></div>
