+++ src/main/resources/templates/sources/sources-add.html
@@ -6,7 +6,7 @@
 
     <link rel="stylesheet"
           href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css">
-    <script src="../../static/js/script.js" defer></script>
+    <script src="../../static/js/scriptGame.js" defer></script>
 </head>
 <body>
 <div class="container">
